Move from custom build system to rollup + closure compiler
