+++ package.json
@@ -11,27 +11,22 @@
   "author": "The AMP HTML Authors",
   "license": "Apache-2.0",
   "scripts": {
-    "build": "tsc -p tsconfig-es6.json",
+    "build": "rollup -c",
     "build-watch": "npm run build -- --watch",
     "clean": "rm -rf dist/ build/",
-    "compile": "mkdir -p dist && node_modules/google-closure-compiler/cli.js -O ADVANCED --js='compile/index.js' --js='build/**.js' --js='!build/**/test-*.js' --js='!build/src/testing/**' --externs='compile/externs.js' --rewrite_polyfills=false > build/closure-output.js",
-    "create-mjs": "cp build/closure-output.js dist/animations.mjs && sed -i '' 's;window.\\([^=]*=\\);export const \\1;g' dist/animations.mjs",
-    "create-cjs": "cp build/closure-output.js dist/animations.js  && sed -i '' 's;window.\\([^=]*=\\);exports.\\1;g' dist/animations.js",
-    "create-global": "npm run tsickle && npm run compile",
     "demo": "npm run build && http-server | grep -m 1 'http:\\.*' | xargs -I{} opener {}/docs/demo",
-    "dist": "npm run clean && npm run create-global && npm run create-mjs && npm run create-cjs && npm run build",
+    "dist": "npm run clean && npm run build",
     "karma": "karma start --browsers Chrome,Firefox karma.conf.js",
     "test": "npm run karma -- --single-run",
     "test-watch": "npm run karma -- --auto-watch",
-    "tsickle": "node_modules/tsickle/src/main.js",
     "prepublish": "npm run dist"
   },
   "devDependencies": {
-    "@types/chai": "^4.1.6",
-    "@types/mocha": "^5.2.5",
-    "chai": "^4.2.0",
-    "google-closure-compiler": "^20181008.0.0",
-    "http-server": "^0.11.1",
+    "@ampproject/rollup-plugin-closure-compiler": "0.12.2",
+    "@types/chai": "4.2.5",
+    "@types/mocha": "5.2.7",
+    "chai": "4.2.0",
+    "http-server": "0.11.1",
     "karma": "^3.0.0",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^2.2.0",
@@ -41,8 +36,13 @@
     "karma-typescript": "^3.0.13",
     "mocha": "^5.2.0",
     "opener": "^1.5.1",
-    "tsickle": "^0.33.0",
-    "typescript": "git://github.com/mprobst/TypeScript.git#5456479a4a"
+    "rollup": "1.25.1",
+    "rollup-plugin-commonjs": "10.1.0",
+    "rollup-plugin-node-resolve": "5.2.0",
+    "rollup-plugin-typescript": "1.0.1",
+    "rollup-plugin-terser": "5.1.2",
+    "typescript": "3.7.2",
+    "tslib": "1.10.0"
   },
   "dependencies": {},
   "files": [
